wholesale_only?
Add a checkbox to product pages that marks certain products
wholesale_only and hides them from normal users.

This only does the hiding on products#show and taxons#show, so if there
are other places you are filtering products by Spree::Product.available
they will not be aware of this property.

https://app.asana.com/0/8134711427023/669808130880462/f